Rename to copy-and-blit, move unit tests
